% File src/library/EpidemicSimulator/man/startSimulator.Rd
\name{startSimulator}
\alias{startSimulator}
\title{Start the Web Simulator}
\description{
  Starts the web shiny application and runs the simulator.
}
\usage{
startSimulator()
}
\arguments{
  \item{optUI}{options for the UI.}
  \item{...}{not used yet.}
}
\seealso{
  \code{\link{getOptionsUI}}.
}

\examples{
# start the simulator
startSimulator()
}

\concept{app}
\concept{web}